refactor(sdk): move concurrency util from token/ to utils/ [SDK-85]
